Update Functionality
The name of some variables was changed.

The syntaxis of the functions whas replaced by arrow functions.

The logic of the pushing of buttons in the buttons array was simplified.

Previous 'for' loops were replaced to 'for each' loops.

It was added the 'atKeyPressed' function that prevents non valid characters to be displayed.

It was added the 'convertDegrees' function and linked to 'degrees' button at click.